requirements.txt
fastmcp
openai
pandas
python-dotenv
requests